requirements.txt
azure-functions==1.20.0
azure-functions-durable>=1.2.2
